requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
faiss-cpu
google-generativeai
langchain-community
langchain-huggingface
python-multipart
websockets
psycopg2-binary
sqlalchemy
python-dotenv
sentence-transformers
